refactor: test case for material request and stock reservation
