+++ .github/workflows/test.yml
@@ -131,13 +131,6 @@ jobs:
           bench --site test_site run-tests \
             --module jewellery_erpnext.jewellery_erpnext.tests.test_material_request_transfer_se
 
-      - name: Run Material Request Transfer
-        run: |
-          cd ~/frappe-bench
-
-          bench --site test_site run-tests \
-            --module jewellery_erpnext.jewellery_erpnext.tests.test_material_request_transfer_type
-
       - name: Run Manufacturing Work Order
         run: |
           cd ~/frappe-bench
@@ -221,13 +214,6 @@ jobs:
 
           bench --site test_site run-tests \
             --module jewellery_erpnext.jewellery_erpnext.tests.test_stock_reservation_entry_mwo
-
-      - name: Run Stock Reservation Entry
-        run: |
-          cd ~/frappe-bench
-
-          bench --site test_site run-tests \
-            --module jewellery_erpnext.jewellery_erpnext.tests.test_stock_reservation_entry_override
 
       - name: Run Repack
         run: |
